+++ src/app/job-status/height-progress/height-progress.component.scss
