+++ src/app/main-screen/main-menu/main-menu.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-main-menu',
+  templateUrl: './main-menu.component.html',
+  styleUrls: ['./main-menu.component.scss'],
+})
+export class MainMenuComponent {
+  public constructor() {}
+}
