+++ src/app/main-screen/printer-status/printer-status.component.scss
